Separate visual and app layer in orders
