+++ src/orders/types.ts
@@ -1,10 +1,4 @@
 import { FilterOpts, MinMax } from "@saleor/types";
-import { OrderStatusFilter } from "@saleor/types/globalTypes";
-
-export interface OrderListFilterOpts {
-  created: FilterOpts<MinMax>;
-  status: FilterOpts<OrderStatusFilter[]>;
-}
 
 export interface OrderDraftListFilterOpts {
   created: FilterOpts<MinMax>;
